file/requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
numpy==1.26.4
scipy==1.11.3
PyYAML==6.0.1

# Deep Learning & AI
torch==2.1.0
torchvision==0.16.0
mediapipe==0.10.14
retina-face==0.0.13   # <-- fixed (PyTorch/TF hybrid, no TF 2.5 lock)

# Audio Processing
pyaudio==0.2.13
librosa==0.10.1
soundfile==0.12.1

# Computer Vision & Image Processing
Pillow==10.1.0
scikit-image==0.22.0

# Data Processing & Analysis
pandas==2.1.2
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0

# Utility Libraries
tqdm==4.66.1
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
